2896 - Add missing id for aria-controls
A recent accessibility audit completed in December 2019, revealed
that our accordion component was using aria-controls attribut on the
accordion section headings button but no elements on the page had an id
that matched the aria-controls values.

Aria-controls attribute is used by assistive tech to help users understand
what they are interacting with and what elements are responsible for various
actions.

I used govuk-frontend accordion as a reference for this fix. There are
no visual differences but the html should be reviewed to see the changes.

Buttons aria-control value matches the accordion content section for
the specific section it relates too.