2968 - Initial markup and styling of the cookie banner
This is the initial markup and styling need to support our cookies
feature. The banner should not be shown just yet as there is some
css that hides it by default. There will be some javascript added
later that will determine whether or not to show the banner.

The see the banner:\

1. Visit any page
2. Open the dev tools
3. Remove `app-cookie-banner--hidden` class from the cookie banner

alternatively,

1. edit `app/views/layout/_cookie-banner.html.erb
2. Remove `app-cookie-banner--hidden` css class on the first line