2968 - Change how we load in Google Analytics
We used to automatically add in Google Analytics (GA) but now
with the introduction of the cookie laws, we have to ask users
for consent before adding third party cookies to their devices.
At the same time we should only inject GA if the user consents
otherwise they are downloading more javascript that will not be
used or needed.

This commit removes GTM (we decided to not replace GA with GTM).
GA should still load as before but we will add the cookie policy
control to this to ensure it only gets used if the user consents
to cookies