[2963] Use subject_code
The teacher-training-api endpoint filters on `subject_code` not the
rails DB id. Update the subject filter to use subject code instead. This
is still only happening internally as we need to interact with the C#
results page for now. Once we have deployed the rails results page we
can update the parameters and remove the C# translation.