[2968] Refactors tracking events
We need to centralise the call to trigger Google Analtics tracking so
we can easily control when to push events to GA. The only difference I
found between the calls `ga` where that the majority were passing an
extra option called `nonInteraction`. I decided to add an extra param
onto analytics.js  `triggerAnalyticsEvent` to include this extra param.
It meant that we could import it into the other files and completely
remove th other `triggerAnalyticsEvent`. We now have one
`triggerAnalyticsEvent` ❤️🎉