[3070] - fix user journey when using the wizard
This fixes a bug where users starting out from the start page
(i.e location page) would be taken straight to the results page if they
search by postcode,town or city. The preferred route was suppose to be
that all location based seached (i.e "search by postcode, town or city"
or "Across the whole of England") should always take the user to the
subjects page to select specific subjects they are interested in.

Provider searchs on the other hand should take users straight to the
results page.

As the location page (ie "Start page") has a dual purpose if they user
has already searched for some courses but decided to edit the location
then they should not have to go back to subject page before returning
to the results page.

I found LocationController create method very hard to read because of
all the redirecting and various params being passed around. I understand
that the reason the params are so complicated is because we have to support
a c# version of the service. We should look to review this in more detail
once we are using rails completely.

I made a start at refactoring the method to attempt to make it easier to
read and understand by moving some of the logic out into private methods.