Don't pass `qualification=none`
Previously we were adding a 'none' value to the filter querystring if
the user deselected all check boxes. When validation failed this was
being written into the back link and caused issues on results page.

This commit removes it and alters the behaviour of the filter page
slightly. If the page is visited with no querystring paramaters then all
checkboxes will remain unchecked. This is more 'obvious' behaviour. The
behaviour when the user gets to the filter page via the results page
remains unchanged.