[3125] Fix location filter backlink
Store the values from the last succesful results page load as hidden
fields in the location filter with a `prev_` prefix.

In the event of an invalid form submission use these to render the
backlink rather than the new invalid parameters.

Also update the links on the non-javascript provider interstitial page
to ensure that they are removed from the querystring before the results
page is loaded.