[3155] Remove elements with duplicate id
I discovered there where elements on the page had the same id.
THis causes a couple of issues

- Elements on a page should have unique ID
- Field labels did not work because label did not know which element
  to focus on (Accessibility issue)

What was causing this is that we output a bunch of hidden fields that
represent params which we track throughout the users journey. We should
have excluded any form fields that were being used on the page.