[3155] Associate location/provider labels & inputs
fixes an accessibility issue with location/provider text fields
not having labels. Screenreaders would not be able to provide the context
of the input to the user.

The labels were present but they were not using the  correct input id
for their `for` attribute.

(No visual difference but users should be able to click on the labels)