[3155] Improve usability of study type
The user does not get the opportunity to select which type of studies they
are interested in as part of the initial search. We include all types
of study by default. The user can change the filter to narrow their search.

We have a validation that ensures at least one checkbox is ticked. However,
if the user decides to untick each option and submits the page, we provide
them with an error message stating they should pick at least one option
but both checkboxes are ticked by default so it maybe confusing to the user
as they did not tick each option and both checkboxes are already ticked.
They maybe confused as to how to correct this error as it seems to have
corrected itself.