[3160] Exclude pagination params from filter_parameters
This resolves the following bug:

Theres a large result set shown on the results page and the user has
used pagination to look through the results. They decide to edit some
filters. After changes to the filter the result set is narrowed down but
the results page tries to take them back to the last set pagination.
This is likely to cause confusion for users or raise errors for us to
investigate. For example the result set is only has 3 courses and the
user was on the 6th page of results.