Don't send unactionable errors to Sentry
`JsonApiClient::Errors::ConnectionError` is caused by the API being
down or unreachable. This doesn't require an action from developers. To
avoid error blindness and reduce the number of messages in the Slack
channel, this filters out all the errors. Our existing monitoring will
pick up if the site is down for an extended period of time.