Fix validations on start wizard when nearing end of cycle
- When 'Settings.cycle_ending_soon' is set to true the
LocationsController was
not redirecting back to the 'start/location' page when validation errors
occur. Instead it was redirecting to the '/cycle-ending-soon' page.
This is also causing Find end to end tests to fail.
- New logic has been added to the LocationsController to detect whether
'Settings.cycle_ending_soon' is true and if so, will redirect back to
'/start/location' when a search validation error occurs.