revert duration/duration_ms formatting for logstash
Turns out this approach was not compatible with the standard way of logging and
can result in breaking the indexing for the 'duration' field, breaking
aggregation and other operations reliant on the 'duration' field. The correct
approach is to leave 'duration_ms', fix any discrepencies in the logit filters
to ensure that it's indexed as a number, and then use 'duration_ms' for
aggregation.