Add feature_flags section to settings.yml
Begin standardising how we handle feature flags by adding a new key to
the settings file and moving all flags here.

This commit also removes the environment variable overrides in
app_config.yml so that we have a single definition for the flags across
all environments. Temporary overrides for any of the environments can be
added to app_config.yml as required.