Updating view partial to use standardised content
As part of a content overhaul on the find application we are updating the copy to
be more in oine with the styling on apply. This PR mirrors the design prototype.
This is a pretty granular solution as I have had limited success in finding the required
info on the TTAPI as a response. Would be intersted if the copy needs to beupdated there
as opposed to here.