Attempt to use Accordion component for subject selection
Had to change the `subject_area_is_selected?` helper - I'm not sure how
this worked with the old markup either because it was comparing ids with
subject codes.

Also lost a few `data-qa` attributes (because they are not part of the
govuk-components markup). This required a few tweaks to the SitePrism
page selectors - these are now using a selector based on an existing CSS class rather than a
`data-qa` attribute.