Use grey regular text for summary labels instead of bold black
The labels within the results list (Course, Financial support, Locations, Vacancies)  are repeated for every item. With the exception of Vacancies (which displays only Yes or No), they're also pretty self-explanatory if you read the values first.

Because of this, we can de-emphasis them by dropping the bold and using dark grey text, to give more focus to the content which does change for each course.