Add GovUk Frontend Cookie Banner Component, remove partial and update route
In order to add more clarity to the cookie banner we need to add the ability to reject cookies. This commit adds a component to
the front end and Javascript to capture the Reject All Cookies button. Also route for cookies page has been updated to reflect
URL on Apply.