Copy the Apply Rubocop config into this repo using Rake
Rubocop's inherit_from doesn't work when following relative includes with
dots in the path from GitHub (as feature in the Apply Rubocop config) because
GitHub returns 302 redirects for those files, with the host missing from
the Location header. And Net::HTTP, which Rubocop uses to make its
requests, doesn't know to reuse the original domain to follow the
redirect, so it crashes.

Fix that, and give ourselves some control over when we pull the latest from
Apply, by providing a Rake task to copy the config over whenever we’re
ready to catch up. Pull the govuk_* cops too so we keep up to date with
those.

This also means we don’t need a network connection to run Rubocop.