Add temporary caching of course API requests
We can't cache course details for any significant length of time in
Find, because there's a requirement for Providers to be able to quickly
update course details and issue corrections. However, we probably should
cache these API calls anyway during the opening week of the cycle, to
help the service handle the expected level of load.

Do this via the following:

- Add a feature flag for this caching behaviour
- Use Faraday middleware to write to a cache when making requests in the
  Course resource
- Give cache entries written here a 10 minute expiry
- Make the Course connection use a valid cache object if the feature
  flag is on, otherwise use a null cache

This will serve requests for the same url, within the expiry, from the
cache.

Alongside these changes, we configure as follows:

- Clear the cache between tests
- Switch caching on in development, test, and qa
- Switch caching off in production (for now)
- Set production explicitly to a null store for now, changing this to a
  Redis store once an instance is configured.