Add feature flag for provider autocomplete
The provider autocomplete on the start page sends requests to TTAPI as
the user types. This gets a bit heavy-going when traffic spikes at start
of cycle.

We're planning on reimplementing it so it uses a locally-stored
providers list for its suggestions. Until that's done, we want to be
able to switch the autocomplete off during peak load.