Use debounce instead of throttle in autocomplete JS
This reduces the number of API requests made (including partially-typed
queries), by waiting for the user to stop typing before starting the
specified wait, followed by triggering the passed in function. In
contract, the throttle function seems to merely ensure spacing between
function calls, meaning the server request can be triggered in the
middle of typing something.

This should have the knock on effect of giving us better quality cache
keys when we store the results of the query.