Use email address as link text for mailto
This makes it clearer that the link will open an email client, which might otherwise be unexpected and annoying, for example if the link opens Outlook but you use Gmail.

Using the email address as link text makes it easier for users to copy and paste the email address into their preferred email app if needed.