Use job environment to update deployment status
The bobheadxi/deployments GitHub Action is used to update deployment
statuses, this is ineffective when deploying to persistent environments
such as qa or prod.  This change removes that action from non-review app
deployment workflows and uses the built in functionality instead.
Review apps continue to use the bobheadxi/deployments action.